[Feat, Refactor] Storyboard 네비게이션 연결, 관련 메서드 수정 등
