+++ JuiceMaker/JuiceMaker/Controller/AdjustStockViewController.swift
@@ -1,0 +1,16 @@
+//
+//  AdjustStockViewController.swift
+//  JuiceMaker
+//
+//  Created by Swain Yun on 2023/09/14.
+//
+
+import UIKit
+
+class AdjustStockViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+
+}
